a function that prints half of a string, followed by a new line.

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -11,7 +11,7 @@
 
 void puts_half(char *str)
 {
-int i, n, len = 0;
+long int i, n, len = 0;
 while (str[len] != '\0')
 len++;
 if (len % 2 == 0)
